refactor(f6-ui): 文件名标准化
